moved to OmniSharp.Roslyn and renamed to CompilationOptionsHelper (less coupling)
